MaruFactory refactoring
