+++ app/build.gradle
@@ -46,11 +46,8 @@ dependencies {
   implementation("org.hyperledger.besu.internal:besu-metrics-core")
   implementation("org.hyperledger.besu.internal:besu-crypto-algorithms")
   implementation("tech.pegasys.teku.internal:p2p")
-  implementation("io.libp2p:jvm-libp2p")
 
   integrationTestImplementation(project(":jvm-libs:test-utils"))
-  integrationTestImplementation(testFixtures(project(":core")))
-  integrationTestImplementation(testFixtures(project(":config")))
   integrationTestImplementation(project(":crypto"))
   integrationTestImplementation "org.hyperledger.besu.internal:besu-ethereum-api"
   integrationTestImplementation(group: 'org.hyperledger.besu.internal', name: 'besu-ethereum-core', classifier: 'test-support')
