+++ app/src/integrationTest/kotlin/maru/app/MaruConsensusSwitchTest.kt
@@ -12,11 +12,11 @@ import awaitTillMaruHasPeers
 import java.io.File
 import maru.testutils.Checks.getMinedBlocks
 import maru.testutils.Checks.verifyBlockTimeWithAGapOn
-import maru.testutils.MaruFactory
 import maru.testutils.besu.BesuFactory
 import maru.testutils.besu.BesuTransactionsHelper
 import maru.testutils.besu.ethGetBlockByNumber
 import maru.testutils.besu.startWithRetry
+import maru.testutils.maru.MaruFactory
 import org.apache.logging.log4j.LogManager
 import org.assertj.core.api.Assertions.assertThat
 import org.hyperledger.besu.tests.acceptance.dsl.blockchain.Amount
