+++ app/src/integrationTest/kotlin/maru/app/MaruLineaFinalizationTest.kt
@@ -15,11 +15,11 @@ import kotlin.time.toJavaDuration
 import linea.domain.BlockParameter
 import linea.ethapi.EthApiClient
 import linea.web3j.ethapi.createEthApiClient
-import maru.testutils.MaruFactory
 import maru.testutils.PeeringNetworkParticipantStack
 import maru.testutils.besu.BesuFactory
 import maru.testutils.besu.BesuTransactionsHelper
 import maru.testutils.besu.ethGetBlockByNumber
+import maru.testutils.maru.MaruFactory
 import org.apache.logging.log4j.LogManager
 import org.assertj.core.api.Assertions.assertThat
 import org.awaitility.kotlin.await
