+++ config/build.gradle
@@ -15,14 +15,6 @@
 
 plugins {
   id 'maru.kotlin-library-conventions'
-  id 'java-test-fixtures'
-}
-
-sourceSets {
-  testFixtures {
-    compileClasspath += sourceSets.main.compileClasspath
-    runtimeClasspath += sourceSets.main.runtimeClasspath
-  }
 }
 
 dependencies {
